refactored settings forms
