+++ client/src/pages/SignIn.js
@@ -27,7 +27,7 @@ const SignIn = () => {
     }
   }, [isAuthenticated, navigate]);
 
-  async function handleFormSubmit(e) {
+  const handleFormSubmit = async (e) => {
     e.preventDefault();
     setIsLoading(true);
     setError(false);
@@ -38,9 +38,9 @@ const SignIn = () => {
       setError(err?.message || "Failed to log in");
     }
     setIsLoading(false);
-  }
+  };
 
-  async function handleGoogleButtonClick(e) {
+  const handleGoogleButtonClick = async (e) => {
     e.preventDefault();
     try {
       dispatch(googleSignInAction());
@@ -48,9 +48,9 @@ const SignIn = () => {
       console.log(err);
       setError("Failed to log in");
     }
-  }
+  };
 
-  async function handleFbButtonClick(e) {
+  const handleFbButtonClick = async (e) => {
     e.preventDefault();
     try {
       dispatch(fbSignInAction());
@@ -58,7 +58,7 @@ const SignIn = () => {
       console.log(err);
       setError("Failed to log in");
     }
-  }
+  };
 
   return (
     <div className="d-flex align-items-center justify-content-center py-5 h-100">
